# openapi-generator/openapi-spec.yaml
openapi: 3.1.0
info:
  title: Scriptos API
  description: The API Specification for the Scriptos Backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /register:
    post:
      operationId: userRegister
      summary: Register a new Scriptos user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/registerRequest"
      responses:
        200:
          description: OK

  /login:
    post:
      operationId: userLogin
      summary: Login as an existing Scriptos user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/loginRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokenResponse"

  /secured:
    get:
      operationId: securedRoute
      summary: A sample secured route for testing token validity
      tags:
        - Security
      responses:
        200:
          description: OK
        403:
          description: Forbidden

  /group:
    post:
      operationId: createGroup
      summary: Create a new Scriptos group
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createGroupRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupIdResponse"
    get:
      operationId: retrieveGroups
      summary: Retrieve all groups viewable by the user
      tags:
        - Groups
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groupListingResponse"


  /group/user:
    post:
      operationId: addUserToGroup
      summary: Add a user to an existing Scriptos group
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addUserToGroupRequest"
      responses:
        200:
          description: OK
    delete:
      operationId: deleteUserFromGroup
      summary: Delete a user from a group they are currently in
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/removeUserFromGroupRequest"
      responses:
        200:
          description: OK

  /group/leave:
    post:
      operationId: leaveGroup
      summary: Leave a group as the logged-in user
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/leaveGroupRequest"
      responses:
        200:
          description: OK

components:
  schemas:
    registerRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    loginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    tokenResponse:
      type: object
      properties:
        token:
          type: string

    errorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string

    createGroupRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string

    groupIdResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid

    addUserToGroupRequest:
      type: object
      required:
        - username
        - groupName
      properties:
        username:
          type: string
        groupName:
          type: string

    removeUserFromGroupRequest:
      type: object
      required:
        - username
        - groupName
      properties:
        username:
          type: string
        groupName:
          type: string

    leaveGroupRequest:
      type: object
      required:
        - groupName
      properties:
        groupName:
          type: string

    groupListingResponse:
      type: array
      items:
        type: object
        required:
          - name
          - description
          - admin
        properties:
          name:
            type: string
          description:
            type: string
          admin:
            type: boolean